Security - Ajout de contrôles au différent point d'insertion de données pour un film - Ajout d'un objet métier pour vérifier les données saisie par l'utilisateur - Ajout d'un objet erreur pour les services

+++ BusinessLogicLayer/Services/moviesServices.js
@@ -1,26 +1,41 @@
 const BaseService = require("./BaseServices");
-const helper = require('../Helpers/helper')
-
+const helper = require('../Helpers/helper');
+const ErrorResponseServices = require("./ResponsesServices/ErrorResponseServices");
+const BoMovie = require('./BusinessObjects/BoMovie')
 class MoviesServices extends BaseService{
     constructor(){
         super()
     }
 
     async createMovieAsync(req,body) {
-        let data = {
-            name: body.name,
-            description: body.description,
-            date: body.date,
-            categoriesId: body.categoriesId
+        let categoriesId = body.categoriesId
+
+        let boMovie = new BoMovie({id: undefined, name: body.name, description: body.description, date: body.date})
+        let ckecking = boMovie.checkData({ignoreId:true})
+
+        if(!ckecking.ok){
+            this.response.setStatus(false)
+            this.response.setError(new ErrorResponseServices(ckecking.error,401).toPrototype())
+            return this.response.toPrototype()
         }
 
-        let result = await req.repositories.getCategoriRepository().existsRange(data.categoriesId)
+        if(categoriesId == undefined || categoriesId.length <= 0){
+            this.response.setStatus(false)
+            this.response.setError(new ErrorResponseServices("un film doit avoir au moins 1 catégorie",401).toPrototype())
+            return this.response.toPrototype()
+        }
+
+        let checkingValue = boMovie.toPrototype()
+        let result = await req.repositories.getCategoriRepository().existsRange(categoriesId)
         
         if(result.success){
-            result = await req.repositories.getMovieRepository().createMovieWithCategorieAsync({data:data,tracking:false});
+            result = await req.repositories.getMovieRepository().createMovieWithCategorieAsync({data:{
+                name:checkingValue.name, description:checkingValue.description,
+                date:checkingValue.date, categoriesId: categoriesId},tracking:false});
         }
 
         this.response.setStatus(result.success)
+
         if(result.success){
             this.response.setData(result.entity)
         }else{
@@ -66,28 +81,65 @@ class MoviesServices extends BaseService{
     }
 
     async updateMovieById(req,id,body){
-        const data = {
-            name: body.name,
-            description: body.description,
-            date: body.date,
-            categoriesId: body.categoriesId
+
+        let categoriesId = body.categoriesId
+        console.log(id)
+
+        let boMovie = new BoMovie({id: id, name: body.name, description: body.description, date: body.date})
+        let ckecking = boMovie.checkData({ignoreId:false})
+        if(!ckecking.ok){
+            this.response.setStatus(false)
+            this.response.setError(new ErrorResponseServices(ckecking.error,401))
+            return this.response.toPrototype()
         }
-        let result = await req.repositories.getCategoriRepository().existsRange(data.categoriesId)
-        if(result.success){
-            result = await req.repositories.getMovieRepository().updateMovieAsync({id:id,data:data,tracking:false})
+        console.log(categoriesId)
+        if(categoriesId == undefined || categoriesId.length <= 0){
+            this.response.setStatus(false)
+            this.response.setError(new ErrorResponseServices("un film doit avoir au moins 1 catégorie",401))
+            return this.response.toPrototype() 
         }
-        
-        this.response.setStatus(result.success)
-        
-        if(result.success){
-            this.response.setData(result.entity)
-        }else{
-            this.response.setError(result.error)
+
+        let checkingValue = boMovie.toPrototype()
+
+        if(this.response.success){
+            
+            let movieExist = await req.repositories.getMovieRepository().existsByIdAsync(id)
+
+            if(!movieExist){
+                this.response.setStatus(false)
+                this.response.setError(new ErrorResponseServices("le film n'existe pas", 404))
+                return this.response.toPrototype()
+            }
+
+            let result = await req.repositories.getCategoriRepository().existsRange(categoriesId)
+            
+            if(result.success){
+                result = await req.repositories.getMovieRepository().updateMovieAsync(
+                    {id:id,data:{
+                    name:checkingValue.name, description:checkingValue.description,
+                    date:checkingValue.date, categoriesId: categoriesId}
+                    ,tracking:false})
+            }
+
+            this.response.setStatus(result.success)
+            if(result.success){
+                this.response.setData(result.entity)
+            }else{
+                this.response.setError(result.error)
+            }
         }
-        return this.response
+
+        return this.response.toPrototype()
     }
 
     async deleteMovieById(req,id){
+
+        if(id === undefined || id <= 0 || isNaN(id)){
+            this.response.setStatus(false)
+            this.response.setError(new ErrorResponseServices("l'identifiant n'existe pas",401))
+            return this.response.toPrototype() 
+        }
+
         this.result = await req.repositories.getMovieRepository().deleteAsync({id:id})
 
         this.response.setStatus(this.result.success)
@@ -98,6 +150,10 @@ class MoviesServices extends BaseService{
             this.response.setError(this.result.error)
         }
         return this.response.toPrototype()
+    }
+
+    _checkingDataMovie(boObject, ignoreId){
+        return 
     }
 }
